+++ jest.config.js
@@ -7,8 +7,10 @@ module.exports = {
     '<rootDir>/src/lib/formatAxiosError.js',
     '<rootDir>/src/lib/openapiValidator.js',
     '<rootDir>/src/lib/jwt.js',
+    '<rootDir>/src/lib/overlays.js',
     '<rootDir>/src/commands/**/*.js',
     '<rootDir>/src/devportal/portal.js',
+    '<rootDir>/src/validate/**/*.js',
   ],
   coverageThreshold: {
     global: {
@@ -35,11 +37,29 @@ module.exports = {
       lines: 90,
       statements: 90,
     },
+    './src/lib/overlays.js': {
+      branches: 90,
+      functions: 100,
+      lines: 90,
+      statements: 90,
+    },
     './src/devportal/portal.js': {
       branches: 90,
       functions: 90,
       lines: 90,
       statements: 90,
+    },
+    './src/validate/**/*.js': {
+      branches: 90,
+      functions: 100,
+      lines: 95,
+      statements: 95,
+    },
+    './src/commands/validate.js': {
+      branches: 90,
+      functions: 100,
+      lines: 95,
+      statements: 95,
     },
   },
 };